プラグインSDKを修正。
- Arrayタイプのパラメータの仕様をもとにリクエストを検証できるように修正。
