+++ dConnectDevicePlugin/dConnectDevicePluginSDK/dconnect-device-plugin-sdk/src/main/java/org/deviceconnect/android/profile/spec/IntegerParameterSpec.java
@@ -51,72 +51,6 @@ public class IntegerParameterSpec extends DConnectParameterSpec<IntegerDataSpec>
         return mDataSpec.isExclusiveMinimum();
     }
 
-    @Override
-    public boolean validate(final Object obj) {
-        if (!super.validate(obj)) {
-            return false;
-        }
-        if (obj == null) {
-            return true;
-        }
-        switch (getFormat()) {
-            case INT32:
-                return validateInt32(obj);
-            case INT64:
-                return validateInt64(obj);
-            default:
-                throw new IllegalStateException();
-        }
-    }
-
-    private boolean validateInt32(final Object param) {
-        if (param instanceof String) {
-            try {
-                return validateRange(Integer.parseInt((String) param));
-            } catch (NumberFormatException e) {
-                return false;
-            }
-        } else if (param instanceof Integer) {
-            return validateRange((Integer) param);
-        } else {
-            return false;
-        }
-    }
-
-    private boolean validateInt64(final Object param) {
-        if (param instanceof String) {
-            try {
-                return validateRange(Long.parseLong((String) param));
-            } catch (NumberFormatException e) {
-                return false;
-            }
-        } else if (param instanceof Long) {
-            return validateRange((Long) param);
-        } else {
-            return false;
-        }
-    }
-
-    private boolean validateRange(final long value) {
-        if (getEnumList() != null) {
-            for (long e : getEnumList()) {
-                if (e == value) {
-                    return true;
-                }
-            }
-            return false;
-        } else {
-            boolean isValid = true;
-            if (getMaximum() != null) {
-                isValid &=  isExclusiveMaximum() ? (getMaximum() > value) : (getMaximum() >= value);
-            }
-            if (getMinimum() != null) {
-                isValid &= isExclusiveMinimum() ? (getMinimum() < value) : (getMinimum() <= value);
-            }
-            return isValid;
-        }
-    }
-
     public static class Builder extends BaseBuilder<Builder> {
 
         private DataFormat mFormat;
