+++ dConnectDevicePlugin/dConnectDevicePluginSDK/dconnect-device-plugin-sdk/src/main/java/org/deviceconnect/android/profile/spec/StringParameterSpec.java
@@ -1,13 +1,9 @@
 package org.deviceconnect.android.profile.spec;
 
 
-import java.util.regex.Pattern;
-
 public class StringParameterSpec extends DConnectParameterSpec<StringDataSpec> {
 
-    private static final Pattern RGB_PATTERN = Pattern.compile("[0-9a-zA-Z]{6}");
-
-    private StringParameterSpec(final DataFormat format) {
+    StringParameterSpec(final DataFormat format) {
         super(new StringDataSpec(format));
     }
 
@@ -37,45 +33,6 @@ public class StringParameterSpec extends DConnectParameterSpec<StringDataSpec> {
 
     public Integer getMaxLength() {
         return mDataSpec.getMaxLength();
-    }
-
-    @Override
-    public boolean validate(final Object obj) {
-        if (!super.validate(obj)) {
-            return false;
-        }
-        if (obj == null) {
-            return true;
-        }
-        if (!(obj instanceof String)) {
-            return false;
-        }
-        String param = (String) obj;
-        switch (getFormat()) {
-            case TEXT:
-                return validateLength(param);
-            case BYTE:
-            case BINARY:
-                return true; // TODO バイナリのサイズ確認(現状、プラグインにはURL形式で通知される)
-            case DATE:
-                return true; // TODO RFC3339形式であることの確認
-            case DATE_TIME:
-                return true; // TODO RFC3339形式であることの確認
-            case RGB:
-                return RGB_PATTERN.matcher(param).matches();
-            default:
-                throw new IllegalStateException();
-        }
-    }
-
-    private boolean validateLength(final String param) {
-        if (getMaxLength() != null && param.length() > getMaxLength()) {
-            return false;
-        }
-        if (getMinLength() != null && param.length() < getMinLength()) {
-            return false;
-        }
-        return true;
     }
 
     public static class Builder extends BaseBuilder<Builder> {
